add ajax functionality for creating and deleting post

+++ controllers/posts_controller.js
@@ -5,10 +5,19 @@ const Comment = require('../models/comment');
 //comment in function is normal code and code in async await code
 module.exports.create = async function (req, res) {
     try {
-        await Post.create({
+        let post = await Post.create({
             content: req.body.content,
             user: req.user._id
         });
+        if (req.xhr) {
+            return res.status(200).json({
+                data: {
+                    post: post,
+                    username: req.user.name
+                },
+                message: 'Post created!!'
+            });
+        }
         req.flash('success', 'post created');
         return res.redirect('back');
     } catch (err) {
@@ -31,7 +40,14 @@ module.exports.create = async function (req, res) {
 module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
-
+        if (req.xhr) {
+            return res.status(200).json({
+                data: {
+                    post: req.params.id
+                },
+                message: 'post deleted'
+            });
+        }
         if (post.user == req.user.id) {
             post.remove();
             await Comment.deleteMany({ post: req.params.id });
